bigquery-analyze-agent/pyproject.toml
[project]
name = "bigquery-analyze-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.13.0",
    "google-cloud-bigquery>=3.25.0",
]
